@@ -6,6 +6,7 @@
     "ng": "ng",
     "build": "npm run env -s && ng build --prod",
     "start": "npm run env -s && ng serve --aot --proxy-config proxy.conf.js",
+    "serve:sw": "npm run build -s && npx http-server ./dist -p 4200",
     "lint": "ng lint --type-check && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
     "test": "ng test",
     "test:ci": "npm run lint -s && ng test --single-run --code-coverage",
@@ -16,40 +17,40 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "^5.2.5",
+    "@angular/common": "^5.2.5",
+    "@angular/compiler": "^5.2.5",
+    "@angular/core": "^5.2.5",
+    "@angular/forms": "^5.2.5",
+    "@angular/http": "^5.2.5",
+    "@angular/platform-browser": "^5.2.5",
+    "@angular/platform-browser-dynamic": "^5.2.5",
+    "@angular/router": "^5.2.5",
     "@ngx-translate/core": "^9.0.1",
-    "@angular/service-worker": "^5.2.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.9",
+    "@angular/service-worker": "^5.2.5",
+    "@ng-bootstrap/ng-bootstrap": "^1.0.0",
     "bootstrap": "4.0.0",
-    "font-awesome": "^4.7.0",
+    "@fortawesome/fontawesome-free-webfonts": "^1.0.4",
     "core-js": "^2.5.2",
     "lodash": "^4.17.4",
     "rxjs": "^5.5.5",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "~1.6.5",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular/cli": "~1.7.2",
+    "@angular/compiler-cli": "^5.2.5",
+    "@angular/language-service": "^5.2.5",
     "@biesbjerg/ngx-translate-extract": "^2.3.1",
-    "@ngx-rocket/scripts": "^2.0.0",
+    "@ngx-rocket/scripts": "^2.1.0",
     "@types/jasmine": "^2.5.52",
     "@types/jasminewd2": "^2.0.2",
-    "@types/lodash": "^4.14.52",
-    "@types/node": "^8.0.10",
+    "@types/lodash": "^4.14.103",
+    "@types/node": "^8.9.4",
     "codelyzer": "^4.1.0",
     "hads": "^1.3.4",
     "htmlhint": "^0.9.13",
     "https-proxy-agent": "^2.0.0",
-    "jasmine-core": "~2.9.1",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "^4.1.0",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "^2.2.0",
@@ -57,14 +58,14 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "^1.1.0",
     "karma-junit-reporter": "^1.2.0",
-    "protractor": "~5.2.2",
+    "protractor": "~5.3.0",
     "puppeteer": "^1.0.0",
-    "stylelint": "~8.4.0",
-    "stylelint-config-recommended-scss": "~3.0.0",
-    "stylelint-config-standard": "~18.0.0",
-    "stylelint-scss": "~2.2.0",
-    "ts-node": "~4.1.0",
+    "stylelint": "~9.1.1",
+    "stylelint-config-recommended-scss": "~3.1.0",
+    "stylelint-config-standard": "~18.1.0",
+    "stylelint-scss": "~2.4.0",
+    "ts-node": "~5.0.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.6.2"
   }
 }
