@@ -4,9 +4,11 @@ import { Routes, RouterModule } from '@angular/router';
 import { Route, extract } from '@app/core';
 import { AboutComponent } from './about.component';
 
-const routes: Routes = Route.withShell([
-  { path: 'about', component: AboutComponent, data: { title: extract('About') } }
-]);
+const routes: Routes = [
+  Route.withShell([
+    { path: 'about', component: AboutComponent, data: { title: extract('About') } }
+  ])
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
