@@ -6,7 +6,7 @@ import { CoreModule } from '@app/core';
 import { AppComponent } from './app.component';
 
 describe('AppComponent', () => {
-  beforeEach(() => {
+  beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [
         RouterTestingModule,
@@ -17,7 +17,7 @@ describe('AppComponent', () => {
       providers: []
     });
     TestBed.compileComponents();
-  });
+  }));
 
   it('should create the app', async(() => {
     const fixture = TestBed.createComponent(AppComponent);
