@@ -1,4 +1,4 @@
-import { Routes } from '@angular/router';
+import { Route as ngRoute, Routes } from '@angular/router';
 
 import { ShellComponent } from './shell/shell.component';
 import { AuthenticationGuard } from './authentication/authentication.guard';
@@ -11,17 +11,17 @@ export class Route {
   /**
    * Creates routes using the shell component and authentication.
    * @param routes The routes to add.
-   * @return {Routes} The new routes using shell as the base.
+   * @return {Route} The new route using shell as the base.
    */
-  static withShell(routes: Routes): Routes {
-    return [{
+  static withShell(routes: Routes): ngRoute {
+    return {
       path: '',
       component: ShellComponent,
       children: routes,
       canActivate: [AuthenticationGuard],
       // Reuse ShellComponent instance when navigating between child views
       data: { reuse: true }
-}];
+    };
   }
 
 }
