@@ -4,10 +4,12 @@ import { Routes, RouterModule } from '@angular/router';
 import { Route, extract } from '@app/core';
 import { HomeComponent } from './home.component';
 
-const routes: Routes = Route.withShell([
-  { path: '', redirectTo: '/home', pathMatch: 'full' },
-  { path: 'home', component: HomeComponent, data: { title: extract('Home') } }
-]);
+const routes: Routes = [
+  Route.withShell([
+    { path: '', redirectTo: '/home', pathMatch: 'full' },
+    { path: 'home', component: HomeComponent, data: { title: extract('Home') } }
+  ])
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
