@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import { of } from 'rxjs/observable/of';
 import { map, catchError } from 'rxjs/operators';
@@ -9,20 +9,21 @@ const routes = {
 };
 
 export interface RandomQuoteContext {
-  // The quote's category: 'nerdy', 'explicit'...
+  // The quote's category: 'dev', 'explicit'...
   category: string;
 }
 
 @Injectable()
 export class QuoteService {
 
-  constructor(private http: Http) { }
+  constructor(private httpClient: HttpClient) { }
 
   getRandomQuote(context: RandomQuoteContext): Observable<string> {
-    return this.http.get(routes.quote(context), { cache: true })
+    return this.httpClient
+      .cache()
+      .get(routes.quote(context))
       .pipe(
-        map((res: Response) => res.json()),
-        map(body => body.value),
+        map((body: any) => body.value),
         catchError(() => of('Error, could not load joke :-('))
       );
   }
